+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](semver).
 - Allow CORS during local development. [#105][#105]
 - Add BUILD_ENV environment variable. [#105][#105]
 - Cache dashboard pledge data client-side. [#104][#104]
+- Add dashboard data export buttons. [#102][#102]
 
 ### Changed
 - Update resource links during local development. [#105][#105]
 - Swap `gulp-connect` for `browser-sync`. [#105][#105]
 - Improve admin messaging.
+- **Breaking change:** Rename `getDashboardMarkup()` to `getSummaryMarkup()`.
 
 ### Deprecated